ViberOutreachManager_Deploy/requirements.txt
fastapi
uvicorn[standard]
streamlit
requests
python-dotenv
